(chfsd) -U accepts double.  Default of rpc last interval is 0.5 seconds

+++ chfsd/fs_hook.h
@@ -1,5 +1,5 @@
 void fs_server_rpc_begin(void *func, const char *diag);
 void fs_server_rpc_end(void *func, const char *diag);
 void fs_server_rpc_wait(void);
-void fs_server_set_rpc_last_interval(int second);
-int fs_server_get_rpc_last_interval(void);
+void fs_server_set_rpc_last_interval(double second);
+double fs_server_get_rpc_last_interval(void);
